fix: resolve muzzle validation failures in kafka instrumentations
Several kafka-clients instrumentations referenced KafkaDecorator (whose
TRACING_ENABLED field now calls Utils.isTracingEnabled) without declaring
kafka_common.Utils as a helper class, causing muzzle to fail resolving
that dependency.

Separately, kafka-streams' StartSpanAdvice/StartSpanAdvice27 advice
classes contained their own private static helper methods, which muzzle
cannot validate since advice classes aren't part of the checked helper
set. Moved that logic into KafkaStreamsDecorator, which is already a
registered helper class.
