+++ dd-java-agent/instrumentation/kafka/kafka-streams-0.11/src/main/java/datadog/trace/instrumentation/kafka_streams/KafkaStreamTaskInstrumentation.java
@@ -3,30 +3,13 @@ package datadog.trace.instrumentation.kafka_streams;
 import static datadog.context.propagation.Propagators.defaultPropagator;
 import static datadog.trace.agent.tooling.InstrumenterModule.TargetSystem.CONTEXT_TRACKING;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
-import static datadog.trace.api.datastreams.DataStreamsContext.create;
-import static datadog.trace.api.datastreams.DataStreamsTags.Direction.INBOUND;
-import static datadog.trace.api.datastreams.DataStreamsTags.createWithGroup;
-import static datadog.trace.bootstrap.instrumentation.api.AgentPropagation.DSM_CONCERN;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activeSpan;
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.traceConfig;
 import static datadog.trace.bootstrap.instrumentation.api.Java8BytecodeBridge.rootContext;
-import static datadog.trace.instrumentation.kafka_common.StreamingContext.STREAMING_CONTEXT;
-import static datadog.trace.instrumentation.kafka_common.Utils.computePayloadSizeBytes;
-import static datadog.trace.instrumentation.kafka_common.Utils.newPathwayOnlySpan;
-import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.BROKER_DECORATE;
 import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.CONSUMER_DECORATE;
-import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.JAVA_KAFKA;
-import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.KAFKA_CONSUME;
-import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.KAFKA_DELIVER;
-import static datadog.trace.instrumentation.kafka_streams.KafkaStreamsDecorator.TIME_IN_QUEUE_ENABLED;
-import static datadog.trace.instrumentation.kafka_streams.ProcessorRecordContextSetter.PR_SETTER;
 import static datadog.trace.instrumentation.kafka_streams.ProcessorRecordContextVisitor.PR_GETTER;
-import static datadog.trace.instrumentation.kafka_streams.StampedRecordContextSetter.SR_SETTER;
 import static datadog.trace.instrumentation.kafka_streams.StampedRecordContextVisitor.SR_GETTER;
 import static java.util.Collections.singletonMap;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
@@ -35,23 +18,17 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
 import datadog.context.ContextScope;
-import datadog.context.propagation.Propagator;
-import datadog.context.propagation.Propagators;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.agent.tooling.annotation.AppliesOn;
 import datadog.trace.api.Config;
-import datadog.trace.api.datastreams.DataStreamsContext;
-import datadog.trace.api.datastreams.DataStreamsTags;
 import datadog.trace.bootstrap.InstrumentationContext;
 import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.instrumentation.kafka_clients.TracingIterableDelegator;
 import java.util.Map;
 import net.bytebuddy.asm.Advice;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.processor.internals.ProcessorNode;
 import org.apache.kafka.streams.processor.internals.ProcessorRecordContext;
@@ -285,8 +262,8 @@ public class KafkaStreamTaskInstrumentation extends InstrumenterModule.DataStrea
 
       final AgentSpan span =
           !KafkaStreamsDecorator.TRACING_ENABLED && traceConfig().isDataStreamsEnabled()
-              ? startDsmOnlyPathwaySpan(record, applicationId)
-              : startTracedConsumeSpan(record, node, applicationId);
+              ? KafkaStreamsDecorator.startDsmOnlyPathwaySpan(record, applicationId)
+              : KafkaStreamsDecorator.startTracedConsumeSpan(record, node, applicationId);
 
       AgentScope agentScope = activateSpan(span);
 
@@ -296,80 +273,6 @@ public class KafkaStreamTaskInstrumentation extends InstrumenterModule.DataStrea
       streamTaskContext.setAgentScope(agentScope);
       InstrumentationContext.get(StreamTask.class, StreamTaskContext.class)
           .put(task, streamTaskContext);
-    }
-
-    /**
-     * Creates and activates the real APM consume span (and, when time-in-queue is enabled, its
-     * broker parent), tags it, and reports DSM checkpoints/transactions off of it.
-     */
-    private static AgentSpan startTracedConsumeSpan(
-        final StampedRecord record, final ProcessorNode node, final String applicationId) {
-      AgentSpan span, queueSpan = null;
-      long timeInQueueStart = SR_GETTER.extractTimeInQueueStart(record);
-      if (timeInQueueStart == 0 || !TIME_IN_QUEUE_ENABLED) {
-        span = startSpan(JAVA_KAFKA.toString(), KAFKA_CONSUME);
-      } else {
-        queueSpan =
-            startSpan(
-                JAVA_KAFKA.toString(), KAFKA_DELIVER, MILLISECONDS.toMicros(timeInQueueStart));
-        BROKER_DECORATE.afterStart(queueSpan);
-        BROKER_DECORATE.onTimeInQueue(queueSpan, record);
-        span = startSpan(JAVA_KAFKA.toString(), KAFKA_CONSUME, queueSpan.spanContext());
-        BROKER_DECORATE.beforeFinish(queueSpan);
-        // The queueSpan will be finished after inner span has been activated to ensure that
-        // spans are written out together by TraceStructureWriter when running in strict mode
-      }
-
-      DataStreamsTags tags = createWithGroup("kafka", INBOUND, applicationId, record.topic());
-
-      final long payloadSize =
-          traceConfig().isDataStreamsEnabled() ? computePayloadSizeBytes(record.value) : 0;
-      reportDsmCheckpointOrInject(span, record, tags, payloadSize);
-
-      CONSUMER_DECORATE.afterStart(span);
-      CONSUMER_DECORATE.onConsume(span, record, node);
-      if (null != queueSpan) {
-        queueSpan.finish();
-      }
-      return span;
-    }
-
-    /**
-     * DSM-only mode (tracing disabled for kafka-streams, DSM enabled): never creates a real span,
-     * so no span is ever written to the agent for this integration. Only the pathway
-     * checkpoint/injection happens, carried by a lightweight, never-collected span shim.
-     */
-    private static AgentSpan startDsmOnlyPathwaySpan(
-        final StampedRecord record, final String applicationId) {
-      final AgentSpan localActiveSpan = activeSpan();
-      final AgentSpan span =
-          newPathwayOnlySpan(localActiveSpan == null ? null : localActiveSpan.spanContext());
-
-      DataStreamsTags tags = createWithGroup("kafka", INBOUND, applicationId, record.topic());
-      final long payloadSize = computePayloadSizeBytes(record.value);
-      reportDsmCheckpointOrInject(span, record, tags, payloadSize);
-      return span;
-    }
-
-    /**
-     * Reports a DSM checkpoint for {@code record}'s topic, or - when in a streaming context and
-     * {@code record}'s topic is a source topic - injects the pathway context so it survives leaving
-     * the topology on another instance of the application.
-     */
-    private static void reportDsmCheckpointOrInject(
-        final AgentSpan span,
-        final StampedRecord record,
-        final DataStreamsTags tags,
-        final long payloadSize) {
-      if (STREAMING_CONTEXT.isDisabledForTopic(record.topic())) {
-        AgentTracer.get()
-            .getDataStreamsMonitoring()
-            .setCheckpoint(span, create(tags, record.timestamp, payloadSize));
-      } else if (STREAMING_CONTEXT.isSourceTopic(record.topic())) {
-        Propagator dsmPropagator = Propagators.forConcern(DSM_CONCERN);
-        DataStreamsContext dsmContext = create(tags, record.timestamp, payloadSize);
-        dsmPropagator.inject(span.with(dsmContext), record, SR_SETTER);
-      }
     }
   }
 
@@ -392,8 +295,8 @@ public class KafkaStreamTaskInstrumentation extends InstrumenterModule.DataStrea
 
       final AgentSpan span =
           !KafkaStreamsDecorator.TRACING_ENABLED && traceConfig().isDataStreamsEnabled()
-              ? startDsmOnlyPathwaySpan(record, applicationId)
-              : startTracedConsumeSpan(record, node, applicationId);
+              ? KafkaStreamsDecorator.startDsmOnlyPathwaySpan(record, applicationId)
+              : KafkaStreamsDecorator.startTracedConsumeSpan(record, node, applicationId);
 
       AgentScope agentScope = activateSpan(span);
 
@@ -403,88 +306,6 @@ public class KafkaStreamTaskInstrumentation extends InstrumenterModule.DataStrea
       streamTaskContext.setAgentScope(agentScope);
       InstrumentationContext.get(StreamTask.class, StreamTaskContext.class)
           .put(task, streamTaskContext);
-    }
-
-    private static long payloadSizeBytes(final ProcessorRecordContext record) {
-      // we have to go through Object to get the RecordMetadata here because the class of `record`
-      // only implements it after 2.7 (and this class is only used if v >= 2.7)
-      if ((Object) record instanceof RecordMetadata) { // should always be true
-        RecordMetadata metadata = (RecordMetadata) (Object) record;
-        return metadata.serializedKeySize() + metadata.serializedValueSize();
-      }
-      return 0;
-    }
-
-    /**
-     * Creates and activates the real APM consume span (and, when time-in-queue is enabled, its
-     * broker parent), tags it, and reports DSM checkpoints/transactions off of it.
-     */
-    private static AgentSpan startTracedConsumeSpan(
-        final ProcessorRecordContext record, final ProcessorNode node, final String applicationId) {
-      AgentSpan span, queueSpan = null;
-      long timeInQueueStart = PR_GETTER.extractTimeInQueueStart(record);
-      if (timeInQueueStart == 0 || !TIME_IN_QUEUE_ENABLED) {
-        span = startSpan(JAVA_KAFKA.toString(), KAFKA_CONSUME);
-      } else {
-        queueSpan =
-            startSpan(
-                JAVA_KAFKA.toString(), KAFKA_DELIVER, MILLISECONDS.toMicros(timeInQueueStart));
-        BROKER_DECORATE.afterStart(queueSpan);
-        BROKER_DECORATE.onTimeInQueue(queueSpan, record);
-        span = startSpan(JAVA_KAFKA.toString(), KAFKA_CONSUME, queueSpan.spanContext());
-        BROKER_DECORATE.beforeFinish(queueSpan);
-        // The queueSpan will be finished after inner span has been activated to ensure that
-        // spans are written out together by TraceStructureWriter when running in strict mode
-      }
-
-      DataStreamsTags tags = createWithGroup("kafka", INBOUND, applicationId, record.topic());
-      long payloadSize = payloadSizeBytes(record);
-      reportDsmCheckpointOrInject(span, record, tags, payloadSize);
-
-      CONSUMER_DECORATE.afterStart(span);
-      CONSUMER_DECORATE.onConsume(span, record, node);
-      if (null != queueSpan) {
-        queueSpan.finish();
-      }
-      return span;
-    }
-
-    /**
-     * DSM-only mode (tracing disabled for kafka-streams, DSM enabled): never creates a real span,
-     * so no span is ever written to the agent for this integration. Only the pathway
-     * checkpoint/injection happens, carried by a lightweight, never-collected span shim.
-     */
-    private static AgentSpan startDsmOnlyPathwaySpan(
-        final ProcessorRecordContext record, final String applicationId) {
-      final AgentSpan localActiveSpan = activeSpan();
-      final AgentSpan span =
-          newPathwayOnlySpan(localActiveSpan == null ? null : localActiveSpan.spanContext());
-
-      DataStreamsTags tags = createWithGroup("kafka", INBOUND, applicationId, record.topic());
-      long payloadSize = payloadSizeBytes(record);
-      reportDsmCheckpointOrInject(span, record, tags, payloadSize);
-      return span;
-    }
-
-    /**
-     * Reports a DSM checkpoint for {@code record}'s topic, or - when in a streaming context and
-     * {@code record}'s topic is a source topic - injects the pathway context so it survives leaving
-     * the topology on another instance of the application.
-     */
-    private static void reportDsmCheckpointOrInject(
-        final AgentSpan span,
-        final ProcessorRecordContext record,
-        final DataStreamsTags tags,
-        final long payloadSize) {
-      if (STREAMING_CONTEXT.isDisabledForTopic(record.topic())) {
-        AgentTracer.get()
-            .getDataStreamsMonitoring()
-            .setCheckpoint(span, create(tags, record.timestamp(), payloadSize));
-      } else if (STREAMING_CONTEXT.isSourceTopic(record.topic())) {
-        Propagator dsmPropagator = Propagators.forConcern(DSM_CONCERN);
-        DataStreamsContext dsmContext = create(tags, record.timestamp(), payloadSize);
-        dsmPropagator.inject(span.with(dsmContext), record, PR_SETTER);
-      }
     }
   }
